+++ .gitignore
@@ -1,0 +1,5 @@
+config.json
+package-lock.json
+yarn.lock
+node_modules/
+captchas/*.*
