+++ events/ready.js
@@ -1,0 +1,6 @@
+module.exports = async(client) => {
+    console.log('Ready!');
+    client.user.setActivity(`on the server.`, {
+      type: 'WATCHING'
+    });
+}
