+++ utils/captcha.js
@@ -1,6 +1,6 @@
 const Jimp = require('jimp');
 
-module.exports = async function createCaptcha() {
+module.exports.createCaptcha = async function createCaptcha() {
     const captcha = Math.random().toString(36).slice(2, 8);
     const image = new Jimp(175, 50, 'white');
     const font = await Jimp.loadFont(Jimp.FONT_SANS_32_BLACK);
@@ -9,6 +9,6 @@ module.exports = async function createCaptcha() {
     const textWidth = Jimp.measureText(font, captcha);
     const textHeight = Jimp.measureTextHeight(font, captcha);
     image.print(font, (w/2 - textWidth/2), (h/2 - textHeight/2), captcha);
-    image.write(`${__dirname}/captchas/${captcha}.png`);
+    image.write(`./captchas/${captcha}.png`);
     return captcha;
 }
